Add files via upload
With a global trading price of 88$ per barrel, oil prices have hit an all-time high in the last seven years. Due to rise in oil price the economy are getting affected very badly.

Oil is a product that goes completely in a different direction for a single market event as the oil prices are rarely based on real-time data, instead, it is driven by externalities making our attempt to forecast it even more challenging

As the economy will be highly affected by oil prices our model will help to understand the pattern in prices to help the customers and businesses to make smart decisions.